Complete exercise 02-flex-header by using flex
Add a padding of 8px and center elements horizontally and vertically with flexbox's properties, in .header div.

Remove default left padding of unordered lists.

Add a fixed width to both left and right links, set them to a display of flex and space-between so that they're spaced out evenly within the fixed width.